backend/requirements.txt
langchain==0.0.350
langchain-community==0.0.10
langchain-core==0.1.10
langchain-openai==0.0.2
langchain-anthropic==0.0.1
wikipedia==1.4.0
python-dotenv==1.0.0
pydantic==1.10.13
duckduckgo-search==4.1.1
langchain_google_genai==0.0.5
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.2
typing-extensions==4.9.0
starlette==0.36.3
anyio==4.2.0
click==8.1.7
h11==0.14.0
idna==3.6
sniffio==1.3.0
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
langsmith==0.0.69
pydantic-core==2.14.5